Fix token when sent request

+++ src/modules/admin/views/Properties.vue
@@ -3,7 +3,7 @@ import { ref, onMounted, nextTick, reactive, watch } from 'vue'
 import { X, Save, SquarePen, Plus, Trash2 } from 'lucide-vue-next'
 import { useToast } from 'vue-toastification'
 import debounce from 'lodash/debounce'
-import { csrf, api } from '../lib/axios'
+import api from '@/lib/axios'
 
 const records = ref([])
 const editingRecord = ref(null)
@@ -22,7 +22,7 @@ const queryParams = reactive({
 const fetchData = async (page = 1) => {
   queryParams.page = page
   try {
-    const response = await api.get(`/properties`, {
+    const response = await api.get(`api/admins/properties`, {
       params: queryParams
     })
     console.log(response)
@@ -67,7 +67,7 @@ const saveRecord = async () => {
   if (!editingRecord.value) return
 
   try {
-    const response = await api.put(`/properties/${editingRecord.value.id}`, {
+    const response = await api.put(`api/admins/properties/${editingRecord.value.id}`, {
       name: editingRecord.value.name
     })
 
@@ -90,7 +90,7 @@ const deleteRecord = async (id) => {
   if (!confirm('Are you sure you want to delete this record?')) return
 
   try {
-    await api.delete(`/properties/${id}`)
+    await api.delete(`api/admins/properties/${id}`)
     const isLastItemOnPage = records.value.length === 1 && pagination.value.current_page > 1
     const nextPage = isLastItemOnPage ? pagination.value.current_page - 1 : pagination.value.current_page
     await fetchData(nextPage)
@@ -119,7 +119,7 @@ const cancelCreating = () => {
 
 const createRecord = async () => {
   try {
-    await api.post(`/properties`, {
+    await api.post(`api/admins/properties`, {
       name: creatingRecord.value.name
     })
     creatingRecord.value = ''
